move reconnect behaviour into own module
